automatically backed up by learn

+++ db/migrate/001_create_costumes.rb
@@ -11,10 +11,10 @@
    def change
       create_table :costumes do |t|
         t.string :name
-        t.integer :price
+        t.float :price
         t.text :image_url
         t.string :size
-        t.timestamps
+        t.timestamps 
       end
    end
  end
